Unify Remote Proxy: single window with HTTPS/HTTP/SSH/Telnet selector
Merge the separate Remote Terminal view into Remote Proxy. One menu item opens
one window where the protocol dropdown (HTTPS/HTTP/SSH/Telnet) switches between
the web iframe (http/https) and the xterm terminal (ssh/telnet).
